pyproject.toml
[project]
name = "openorbit"
version = "0.1.0"
description = "OSINT platform for orbital launch intelligence"
readme = "README.md"
authors = [
    { name = "fwillemen", email = "fwillemen@users.noreply.github.com" }
]
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.1",
    "apscheduler>=3.11.2",
    "fastapi>=0.135.1",
    "httpx>=0.28.1",
    "pydantic-settings>=2.13.1",
    "structlog>=25.5.0",
    "uvicorn>=0.42.0",
]

[build-system]
requires = ["uv_build>=0.10.9,<0.11.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.1.0",
    "pyyaml>=6.0.3",
    "respx>=0.22.0",
    "ruff>=0.15.7",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "--cov=src --cov-report=term-missing --cov-report=html"

[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*", "*/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_generics = true
warn_unreachable = true
no_implicit_optional = true

[[tool.mypy.overrides]]
module = "structlog.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "openorbit.main"
disable_error_code = ["arg-type"]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "SIM", # flake8-simplify
]
ignore = [
    "B008",  # Allow function calls in argument defaults (FastAPI dependency injection pattern)
    "E501",  # Line too long — docstrings and descriptions may exceed 88 chars
]

[tool.ruff.lint.isort]
known-first-party = ["openorbit"]
